pyproject.toml
[project]
name = "rest-full-api"
version = "0.1.0"
description = ""
authors = [
    {name = "Ilya",email = "ilya.piltyay@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django (>=5.2.4,<6.0.0)",
    "dotenv (>=0.9.9,<0.10.0)",
    "psycopg2-binary (>=2.9.10,<3.0.0)",
    "python-dotenv (>=1.1.1,<2.0.0)",
    "djangorestframework (>=3.16.0,<4.0.0)",
    "pillow (>=11.3.0,<12.0.0)",
    "redis (>=6.2.0,<7.0.0)",
    "ipython (>=9.4.0,<10.0.0)",
    "django-filter (>=25.1,<26.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.lint.dependencies]
isort = "^6.0.1"
black = "^25.1.0"
mypy = "^1.17.0"


[tool.poetry.group.dev.dependencies]
flake8 = "^7.3.0"

